Added underscore prefixes to most files
The convention for sass is to add an underscore prefix to any files
that are primarily meant to be imported. Also, the rails convention
of adding the ".css" prior to the .scss convention isn't necessary
and is also meant for files that are supposed to be directly
converted to css (rather than be imported).